feat: Add newpipe search support

+++ app/src/main/java/app/suhasdissa/mellowmusic/backend/models/songs/SongItem.kt
@@ -9,7 +9,7 @@ data class SongItem(
     @SerialName("title") var title: String = "",
     @SerialName("thumbnail") var thumbnail: String = "",
     @SerialName("uploaderName") var uploaderName: String = "",
-    @SerialName("duration") var duration: Int = 0
+    @SerialName("duration") var duration: Long = 0
 ) {
     val videoId
         get() = url.replace("/watch?v=", "")
